cmake files and executables changes for win32

+++ apps/scene2pset/scene2pset.cc
@@ -183,7 +183,7 @@ main (int argc, char** argv)
     /* Iterate over views and get points. */
     mve::Scene::ViewList& views(scene->get_views());
 #pragma omp parallel for schedule(dynamic)
-    for (std::size_t i = 0; i < views.size(); ++i)
+    for (std::int64_t i = 0; i < views.size(); ++i)
     {
         mve::View::Ptr view = views[i];
         if (view == nullptr)
