+++ apps/sfmrecon/sfmrecon.cc
@@ -413,7 +413,7 @@ sfm_reconstruct (AppSettings const& conf)
     }
 
 #pragma omp parallel for schedule(dynamic,1)
-    for (std::size_t i = 0; i < bundle_cams.size(); ++i)
+    for (std::int64_t i = 0; i < bundle_cams.size(); ++i)
     {
         mve::View::Ptr view = views[i];
         mve::CameraInfo const& cam = bundle_cams[i];
